requirements.txt
DateTime==4.0.1
Flask==0.10
Flask-Limiter==0.7.5
Jinja2==2.7.3
MarkupSafe==0.23
MicroblogCrawler==1.4.1
Werkzeug==0.10.1
argparse==1.2.1
itsdangerous==0.24
limits==1.0.4
lxml==3.4.4
python-dateutil==2.4.2
pytz==2015.2
requests==2.6.2
six==1.9.0
wsgiref==0.1.2
zope.interface==4.1.2
